Updated time manipulation so it is more intuitive and less prone to errors.
Representation of time is handled by timespan_t. This structure has static method that returns a representation of time from different base unit (from_minutes, from_seconds, from_millis). It also provides method to return the appropriate amount (total_minutes, total_seconds, total_millis).

The max value is the maximum valid time while the min value should represent the minimum valid time. For now, because of legacy time usage, the minimum value is set to -1 second.

The zero value is available as a constant because it is the only value that means the same whether it is minutes, seconds, or milliseconds.

A version of timespan_t using integer time with a resolution of 1 milliseconds can be used by defining SC_USE_INTEGER_TIME. The integer time is a work in progress.

git-svn-id: http://simulationcraft.googlecode.com/svn/trunk@10872 3b4652a1-8050-0410-ac47-3d40261b0f8b