Return of the adds, part I.
It is again possible to spawn adds as a raid event. New option "name" allows the adds to be named. Each add will have a running number suffixed to the name. Another option "master" specifies which target to spawn the adds for (i.e., who becomes the master of the adds). By default, Fluffy_Pillow will control the hordes.

Oh, and a single add raid event cannot spawn overlapping sets of adds, i.e., the minimum cooldown cannot be shorter than the duration of the event.

Also, the introduction of adds (dynamic arise/demise of enemies) required some checking in the core sim systems to make sure that actions are not executed on sleeping things.

OS X:
Add sc_import.cpp to project.



git-svn-id: http://simulationcraft.googlecode.com/svn/trunk@14446 3b4652a1-8050-0410-ac47-3d40261b0f8b