Warrior: - Sweeping Strikes actually does take target modifiers into account, tested on live/ptr today. - Sweeping strikes does not proc from AoE abilities such as bladestorm/dragon roar/thunderclap.
Additional Note: Sweeping strikes has a lot of unusual interactions. It can proc strikes of opportunity on the target that it hits (target2), which then will proc sweeping strikes on target1, which also has a chance to proc opportunity strikes as well, and can continue for quite a while.  

Picture of one regular white swing, nothing else.
http://gyazo.com/bfc1e5c9b5790ceb145b1e12ee15f2f7

I have no clue how to implement that. 

git-svn-id: http://simulationcraft.googlecode.com/svn/trunk@17374 3b4652a1-8050-0410-ac47-3d40261b0f8b