Added buff for cooldown reduction trinkets. - This will allow action lists to easily accommodate the differences in ability cooldowns created by the trinket.
Example: Storm Bolt is used every time it is up without the trinket, but when the trinket is equipped the cd on storm bolt aligns perfectly with colossus smash, so we should always use it there.
/storm_bolt,if=(buff.cooldown_reduction.up&debuff.colossus_smash.up)|buff.cooldown_reduction.down

I'm not sure if this will ever be needed, but using buff.cooldown_reduction.value will give the multiplier used to reduce the cooldown by, as higher ilevels of the trinket reduce cooldowns even more. 

Example: // Will only use storm bolt if a Normal/LFR Cooldown Reduction Trinket is equipped. 
/storm_bolt,if=buff.cooldown_reduction.value>0.7

// Will only use storm bolt if a heroic+ CDR trinket is equipped.
/storm_bolt,if=buff.cooldown_reduction.value<0.7

git-svn-id: http://simulationcraft.googlecode.com/svn/trunk@17393 3b4652a1-8050-0410-ac47-3d40261b0f8b